finished with moving the simulate delete logic to a service file
